Split the Binding Class for template into parials. This is just to make it more readable.
